requirements.txt
transformers[torch]==4.46.2
neo4j==5.26.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36